+++ packages/pyright-internal/src/localization/package.nls.en-us.json
@@ -342,7 +342,8 @@
         "orPatternMissingName": "All subpatterns within an \"or\" pattern must target the same names",
         "overlappingKeywordArgs": "Typed dictionary overlaps with keyword parameter: {names}",
         "overlappingOverload": "Overload {obscured} for \"{name}\" will never be used because its parameters overlap overload {obscuredBy}",
-        "overloadAbstractMismatch": "Overloaded methods must all be abstract or not",
+        "overloadAbstractMismatch": "Overloads must all be abstract or not",
+        "overloadAbstractImplMismatch": "Overloads must match abstract status of implementation",
         "overloadClassMethodInconsistent": "Overloads for \"{name}\" use @classmethod inconsistently",
         "overloadFinalInconsistencyImpl": "Overload for \"{name}\" is marked @final but implementation is not",
         "overloadFinalInconsistencyNoImpl": "Overload {index} for \"{name}\" is marked @final but overload 1 is not",
