+++ packages/pyright-internal/src/tests/samples/abstractClass5.py
@@ -5,24 +5,22 @@ from abc import ABC, abstractmethod
 from typing import Union, overload
 
 
-class Foo(ABC):
+class ClassA(ABC):
     @overload
-    @abstractmethod
     def func1(self, a: int) -> int:
         pass
 
     @overload
     @abstractmethod
+    # This should generate an error because this overload is
+    # missing an abstractmethod overload.
     def func1(self, a: float) -> float:
         pass
 
     @overload
-    @abstractmethod
     def func1(self, a: str) -> str:
         ...
 
-    # This should generate an error because this overload is
-    # missing an abstractmethod overload.
     def func1(self, a: Union[int, float, str]) -> Union[int, float, str]:
         raise NotImplementedError()
 
@@ -32,11 +30,29 @@ class Foo(ABC):
 
     @overload
     @abstractmethod
-    # This should generate an error because this overload has
-    # an abstractmethod overload.
     def func2(self, a: int) -> int:
         pass
 
     @abstractmethod
     def func2(self, a: Union[int, str]) -> Union[int, str]:
         raise NotImplementedError()
+
+    @overload
+    def func3(self, a: str) -> str:  # pyright: ignore[reportNoOverloadImplementation]
+        ...
+
+    @overload
+    @abstractmethod
+    # This should generate an error because the abstract status is inconsistent.
+    def func3(self, a: int) -> int:
+        ...
+
+    @overload
+    @abstractmethod
+    def func4(self, a: str) -> str:  # pyright: ignore[reportNoOverloadImplementation]
+        ...
+
+    @overload
+    # This should generate an error because the abstract status is inconsistent.
+    def func4(self, a: int) -> int:
+        ...
